+++ CHANGELOG.md
@@ -152,3 +152,4 @@
 * Added burst fire to weapons [WIP]
 * Implemented projectils-per-shot and ammo-per-shot for weapons [WIP]
 * Extended projectile and explosion handling for weapons [WIP]
+* Extended effects handling for sprite weapon view models [WIP]